requirements.txt
fastapi           
uvicorn           
tensorflow        
pydantic          
google-cloud-storage  